Module asio::ip

Structs

Icmp
Encapsulates the flags needed for ICMP(v6).

IcmpResolver
The ICMP(v6) resolver type.

IcmpSocket
The ICMP(v6) socket type.

IpAddrV4
Implements IP version 4 style addresses.

IpAddrV6
Implements IP version 6 style addresses.

IpEndpoint
Describes an endpoint for a version-independent IP socket.

LlAddr
Implements Link-Layer addresses.

Passive
A query of the resolver for the passive mode.

ResolveEntry
An entry produced by a resolver.

ResolveIter
An iterator over the entries produced by a resolver.

Tcp
Encapsulates the flags needed for TCP.

TcpListener
The TCP listener type.

TcpResolver
The TCP resolver type.

TcpSocket
The TCP socket type.

Udp
Encapsulates the flags needed for UDP.

UdpResolver
The UDP resolver type.

UdpSocket
The UDP socket type.

V6Only
Socket option for get/set an IPv6 socket supports IPv6 communication only.

Enums

IpAddr
Implements version-independent IP addresses.

Traits

IpSocket
The IP-based socket tag.

ResolveQuery
A query to be passed to a resolver.

Resolver
Provides endpoint resolution functionality.

ToEndpoint
Provides convert to endpoint.

Type Definitions

IcmpEndpoint
The type of a ICMP(v6) endpoint.

TcpEndpoint
The type of a TCP endpoint.

UdpEndpoint
The type of a UDP endpoint.