Struct asio::ip::IpAddrV6

pub struct IpAddrV6 {
    // some fields omitted
}

Implements IP version 6 style addresses.

Methods

impl IpAddrV6

fn new(a: u16, b: u16, c: u16, d: u16, e: u16, f: u16, g: u16, h: u16) -> IpAddrV6

Constructs a IP-v6 address.

The result will represent the IP address a:b:c:d:e:f:g:h

Examples

use asio::ip::IpAddrV6;

let ip = IpAddrV6::new(0,0,0,0,0,0,0,1);

fn with_scope_id(a: u16, b: u16, c: u16, d: u16, e: u16, f: u16, g: u16, h: u16, scope_id: u32) -> IpAddrV6

Constructs a IP-v6 address with set a scope-id.

The result will represent the IP address a:b:c:d:e:f:g:h%[scope-id]

Examples

use asio::ip::IpAddrV6;

let ip = IpAddrV6::with_scope_id(0,0,0,0,0,0,0,1,0x01);

fn any() -> IpAddrV6

Constructs a unspecified IP-v6 address.

Examples

use asio::ip::IpAddrV6;

let ip = IpAddrV6::any();
assert_eq!(ip, IpAddrV6::new(0,0,0,0,0,0,0,0));

fn loopback() -> IpAddrV6

Constructs a loopback IP-v6 address.

Examples

use asio::ip::IpAddrV6;

let ip = IpAddrV6::loopback();
assert_eq!(ip, IpAddrV6::new(0,0,0,0,0,0,0,1));

fn from_bytes(bytes: [u8; 16], scope_id: u32) -> IpAddrV6

Constructs a IP-v6 address from 16-octet bytes.

Examples

use asio::ip::IpAddrV6;

let ip = IpAddrV6::from_bytes([0,1,2,3, 4,5,6,7, 8,9,10,11, 12,13,14,15], 0);
assert_eq!(ip, IpAddrV6::new(0x0001, 0x0203,0x0405,0x0607,0x0809,0x0A0B, 0x0C0D, 0x0E0F));

fn get_scope_id(&self) -> u32

Returns a scope-id.

Examples

use asio::ip::IpAddrV6;

let ip = IpAddrV6::with_scope_id(0,0,0,0,0,0,0,0,0x10);
assert_eq!(ip.get_scope_id(), 16);

fn set_scope_id(&mut self, scope_id: u32)

Sets a scope-id.

Examples

use asio::ip::IpAddrV6;

let mut ip = IpAddrV6::loopback();
assert_eq!(ip.get_scope_id(), 0);

ip.set_scope_id(0x10);
assert_eq!(ip.get_scope_id(), 16);

fn is_unspecified(&self) -> bool

Returns true if this is a unspecified address.

fn is_loopback(&self) -> bool

Returns true if this is a loopback address.

fn is_link_local(&self) -> bool

Returns true if this is a link-local address.

fn is_site_local(&self) -> bool

Returns true if this is a site-local address.

fn is_multicast(&self) -> bool

Returns true if this is a some multicast address.

fn is_multicast_global(&self) -> bool

Returns true if this is a multicast address for global.

fn is_multicast_link_local(&self) -> bool

Returns true if this is a multicast address for link-local.

fn is_multicast_node_local(&self) -> bool

Returns true if this is a multicast address for node-local.

fn is_multicast_org_local(&self) -> bool

Returns true if this is a multicast address for org-local.

fn is_multicast_site_local(&self) -> bool

Returns true if this is a multicast address for site-local.

fn is_v4_mapped(&self) -> bool

Returns true if this is a mapped IP-v4 address.

fn is_v4_compatible(&self) -> bool

Returns true if this is a IP-v4 compatible address.

fn as_bytes(&self) -> &[u8; 16]

Retruns a 16 octets array.

fn to_v4(&self) -> Option<IpAddrV4>

Retruns a IP-v4 address if this is a convertable address.

fn v4_mapped(addr: &IpAddrV4) -> Self

Constructs a mapped IP-v4 address.

Ex. 192.168.0.1 => ::ffff:192.168.0.1

fn v4_compatible(addr: &IpAddrV4) -> Option<Self>

Constructs a IP-v4 compatible address if the addr isn't in 0.0.0.0, 0.0.0.1.

Ex. 192.168.0.1 => ::192.168.0.1